back to stream mode

+++ app.py
@@ -145,7 +145,7 @@ def main(args):
         logging.info(cntr)
         producer.send(args.topic, json.dumps(generate_event(TXN_TS+TXN_INCREMENT,crdNo)).encode(), json.dumps(crdNo).encode())
         producer.send(args.histTopic, json.dumps(generate_event(TXN_TS+TXN_INCREMENT,crdNo)).encode(), json.dumps(crdNo).encode())
-        cntr = cntr +1
+        cntr = (int)cntr +1
         time.sleep(1.0 / 100)
 
 def get_arg(env, default):
